CT-31 Refactor routing and pages styles
