+++ client/web/src/pages/ActivationNeeded/index.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router';
-import cl from './activationsuccess.module.scss';
+import styles from './styles.module.scss';
 import { AuthState } from '../../types/auth';
 import { useTypedSelector } from '../../hooks/useTypedSelector';
 
@@ -10,12 +10,12 @@ const ActivationSuccess = () => {
 
     return (
 
-        <div className={cl['activation_success']} >
+        <div className={styles['activation_success']} >
 
             {authState === AuthState.Loggedin && (
                 <Navigate to="/" replace={true} />
             )}
-            <div className={cl['success-message']}>
+            <div className={styles['success-message']}>
                 <h2>
                     You was successfully registered!!!
                 </h2>
