+++ client/web/src/pages/ActivationSuccess/styles.module.scss
@@ -1,0 +1,16 @@
+.activation_success {
+    flex: 1 1 auto;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-height: 300px;
+    min-width: 300px;
+}
+
+.success-message {
+    display: flex;
+    flex-direction: column;
+    width: 340px;
+    background: #fff;
+    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
+}
