--- client/web/src/pages/Home/home.module.scss
@@ -1,5 +1,0 @@
-.home {
-    flex: 1 1 auto;
-    display: flex;
-    justify-content: space-between;
-}
