+++ client/web/src/pages/Home/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import styles from './styles.module.scss';
+import Sidebar from '../../components/Sidebar/Sidebar';
+import Map from '../../components/Map/Map';
+
+const Home = () => {
+    return (
+        <div className={styles['home']} >
+            <Sidebar></Sidebar>
+            <Map></Map>
+        </div>
+    );
+};
+
+export default Home;
