+++ client/web/src/pages/Home/styles.module.scss
@@ -1,0 +1,5 @@
+.home {
+    flex: 1 1 auto;
+    display: flex;
+    justify-content: space-between;
+}
